requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
pyarrow>=14.0.1
joblib>=1.3.2
scikit-learn>=1.3.2
numpy>=1.24.0
click>=8.1.7